+++ src/rer/pubblicazioni/tests/test_pubblicazione.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from plone import api
 from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
@@ -17,21 +16,20 @@ class PubblicazioneIntegrationTest(unittest.TestCase):
 
     def setUp(self):
         """Custom shared utility setup for tests."""
-        self.portal = self.layer['portal']
-        setRoles(self.portal, TEST_USER_ID, ['Manager'])
-        self.installer = api.portal.get_tool('portal_quickinstaller')
+        self.portal = self.layer["portal"]
+        setRoles(self.portal, TEST_USER_ID, ["Manager"])
 
     def test_schema(self):
-        fti = queryUtility(IDexterityFTI, name='Pubblicazione')
+        fti = queryUtility(IDexterityFTI, name="Pubblicazione")
         schema = fti.lookupSchema()
         self.assertEqual(IPubblicazione, schema)
 
     def test_fti(self):
-        fti = queryUtility(IDexterityFTI, name='Pubblicazione')
+        fti = queryUtility(IDexterityFTI, name="Pubblicazione")
         self.assertTrue(fti)
 
     def test_factory(self):
-        fti = queryUtility(IDexterityFTI, name='Pubblicazione')
+        fti = queryUtility(IDexterityFTI, name="Pubblicazione")
         factory = fti.factory
         obj = createObject(factory)
         self.assertTrue(IPubblicazione.providedBy(obj))
@@ -39,7 +37,7 @@ class PubblicazioneIntegrationTest(unittest.TestCase):
     def test_adding(self):
         obj = api.content.create(
             container=self.portal,
-            type='Pubblicazione',
-            id='Pubblicazione',
+            type="Pubblicazione",
+            id="Pubblicazione",
         )
         self.assertTrue(IPubblicazione.providedBy(obj))